refactor: move video crawl schema to shared schema directory
