+++ src/apps/gateway/videos/routes/fix-videos-thumbnail/index.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { envConfig } from 'src/utils/envConfig';
 import { ValidatedRequest } from 'src/utils/validator';
-import { WebhookRequest } from '../../schema';
+import { HasuraWebhookRequest } from 'src/schema/hasura';
 import { verifySignature } from 'src/services/videos/convert/validator';
 import { AppError, AppResponse } from 'src/utils/schema';
 import { queues } from 'src/utils/systemConfig';
@@ -11,7 +11,7 @@ import { getVideoMissingThumbnail } from 'src/database/queries/videos';
 
 const fixVideosThumbnail = async (req: Request, res: Response) => {
   const { ioServiceUrl } = envConfig;
-  const { validatedData } = req as ValidatedRequest<WebhookRequest>;
+  const { validatedData } = req as ValidatedRequest<HasuraWebhookRequest>;
   const { signatureHeader } = validatedData;
 
   if (!verifySignature(signatureHeader)) {
