+++ src/schema/videos/fix-thumbnail/index.test.ts
